+++ sportorg/modules/live/orgeo.py
@@ -253,6 +253,7 @@ async def create(url, data, race_data, log, *, session):
             "ResultManual",
             "ResultRfidImpinj",
             "ResultSrpid",
+            "ResultHuichang",
         ]:
             person_data = _get_person(item, race_data)
             if person_data:
@@ -296,6 +297,7 @@ async def create_online_cp(url, data, race_data, log, *, session):
             "ResultManual",
             "ResultRfidImpinj",
             "ResultSrpid",
+            "ResultHuichang",
         ]:
             try:
                 res = _get_result_by_id(item, race_data)
@@ -379,6 +381,7 @@ async def delete(url, data, race_data, log, *, session):
             "ResultManual",
             "ResultRfidImpinj",
             "ResultSrpid",
+            "ResultHuichang",
         ]:
             person_data = _get_person(item, race_data)
             if person_data: